Partial grafting of Path and DPath ext. into siepic_tools (all but snap)
