+++ klayout_dot_config/python/siepic_tools/extend/__init__.py
@@ -1,1 +1,2 @@
-import siepic_tools.extend.point_multiplication  # noqa
+import siepic_tools.extend.point  # noqa
+import siepic_tools.extend.path  # noqa
