+++ klayout_dot_config/python/siepic_tools/utils/sampling.py
@@ -3,153 +3,154 @@
 
 # Modified by Thomas Ferreira de Lima @thomaslima
 # License: CC BY 4.0 https://creativecommons.org/licenses/by/4.0/
+try:
+    import numpy as np
+except ImportError:
+    def sample_function(*args, **kwargs):
+        raise NotImplementedError("Need numpy package to be installed")
+else:
+    def sample_function(func, points, tol=0.001, min_points=8, max_level=16,
+                        sample_transform=None):
+        """
+        Sample a 1D function to given tolerance by adaptive subdivision.
 
-import numpy as np
+        The function itself can be multidimensional.
 
+        The result of sampling is a set of points that, if plotted,
+        produces a smooth curve with also sharp features of the function
+        resolved.
 
-def sample_function(func, points, tol=0.001, min_points=8, max_level=16,
-                    sample_transform=None):
-    """
-    Sample a 1D function to given tolerance by adaptive subdivision.
+        Parameters
+        ----------
+        func : callable
+            Function func(x) of a single argument. It is assumed to be vectorized.
+        points : array-like, 1D
+            Initial points to sample, sorted in ascending order.
+            These will determine also the bounds of sampling.
+        tol : float, optional
+            Tolerance to sample to. The condition is roughly that the total
+            length of the curve on the (x, y) plane is computed up to this
+            tolerance.
+        min_point : int, optional
+            Minimum number of points to sample.
+        max_level : int, optional
+            Maximum subdivision depth.
+        sample_transform : callable, optional
+            Function w = g(x, y). The x-samples are generated so that w
+            is sampled.
 
-    The function itself can be multidimensional.
+        Returns
+        -------
+        x : ndarray
+            X-coordinates (WARNING: returns sorted samples)
+        y : ndarray
+            Corresponding values of func(x)
 
-    The result of sampling is a set of points that, if plotted,
-    produces a smooth curve with also sharp features of the function
-    resolved.
+        Notes
+        -----
+        This routine is useful in computing functions that are expensive
+        to compute, and have sharp features --- it makes more sense to
+        adaptively dedicate more sampling points for the sharp features
+        than the smooth parts.
 
-    Parameters
-    ----------
-    func : callable
-        Function func(x) of a single argument. It is assumed to be vectorized.
-    points : array-like, 1D
-        Initial points to sample, sorted in ascending order.
-        These will determine also the bounds of sampling.
-    tol : float, optional
-        Tolerance to sample to. The condition is roughly that the total
-        length of the curve on the (x, y) plane is computed up to this
-        tolerance.
-    min_point : int, optional
-        Minimum number of points to sample.
-    max_level : int, optional
-        Maximum subdivision depth.
-    sample_transform : callable, optional
-        Function w = g(x, y). The x-samples are generated so that w
-        is sampled.
+        Examples
+        --------
+        >>> def func(x):
+        ...     '''Function with a sharp peak on a smooth background'''
+        ...     a = 0.001
+        ...     return x + a**2/(a**2 + x**2)
+        ...
+        >>> x, y = sample_function(func, [-1, 1], tol=1e-3)
 
-    Returns
-    -------
-    x : ndarray
-        X-coordinates (WARNING: returns sorted samples)
-    y : ndarray
-        Corresponding values of func(x)
+        >>> import matplotlib.pyplot as plt
+        >>> xx = np.linspace(-1, 1, 12000)
+        >>> plt.plot(xx, func(xx), '-', x, y[0], '.')
+        >>> plt.show()
 
-    Notes
-    -----
-    This routine is useful in computing functions that are expensive
-    to compute, and have sharp features --- it makes more sense to
-    adaptively dedicate more sampling points for the sharp features
-    than the smooth parts.
+        """
+        with np.errstate(divide='ignore', invalid='ignore'):
+            return _sample_function(func, points, values=None, mask=None, depth=0,
+                                    tol=tol, min_points=min_points, max_level=max_level,
+                                    sample_transform=sample_transform)
 
-    Examples
-    --------
-    >>> def func(x):
-    ...     '''Function with a sharp peak on a smooth background'''
-    ...     a = 0.001
-    ...     return x + a**2/(a**2 + x**2)
-    ...
-    >>> x, y = sample_function(func, [-1, 1], tol=1e-3)
+    def _sample_function(func, points, values=None, mask=None, tol=0.05,
+                         depth=0, min_points=16, max_level=16,
+                         sample_transform=None):
+        points = np.unique(points)
 
-    >>> import matplotlib.pyplot as plt
-    >>> xx = np.linspace(-1, 1, 12000)
-    >>> plt.plot(xx, func(xx), '-', x, y[0], '.')
-    >>> plt.show()
+        if values is None:
+            values = np.atleast_2d(func(points))
 
-    """
-    with np.errstate(divide='ignore', invalid='ignore'):
-        return _sample_function(func, points, values=None, mask=None, depth=0,
-                                tol=tol, min_points=min_points, max_level=max_level,
-                                sample_transform=sample_transform)
+        if mask is None:
+            mask = slice(None)
 
+        if depth > max_level:
+            # recursion limit
+            return points, values
 
-def _sample_function(func, points, values=None, mask=None, tol=0.05,
-                     depth=0, min_points=16, max_level=16,
-                     sample_transform=None):
-    points = np.unique(points)
+        x_a = points[..., :-1][..., mask]
+        x_b = points[..., 1:][..., mask]
 
-    if values is None:
-        values = np.atleast_2d(func(points))
+        x_c = .5 * (x_a + x_b)
+        y_c = np.atleast_2d(func(x_c))
 
-    if mask is None:
-        mask = slice(None)
+        x_2 = np.r_[points, x_c]
+        y_2 = np.r_['-1', values, y_c]
+        j = np.argsort(x_2)
 
-    if depth > max_level:
-        # recursion limit
-        return points, values
+        x_2 = x_2[..., j]
+        y_2 = y_2[..., j]
 
-    x_a = points[..., :-1][..., mask]
-    x_b = points[..., 1:][..., mask]
+        # -- Determine the intervals at which refinement is necessary
 
-    x_c = .5 * (x_a + x_b)
-    y_c = np.atleast_2d(func(x_c))
-
-    x_2 = np.r_[points, x_c]
-    y_2 = np.r_['-1', values, y_c]
-    j = np.argsort(x_2)
-
-    x_2 = x_2[..., j]
-    y_2 = y_2[..., j]
-
-    # -- Determine the intervals at which refinement is necessary
-
-    if len(x_2) < min_points:
-        mask = np.ones([len(x_2) - 1], dtype=bool)
-    else:
-        # represent the data as a path in N dimensions (scaled to unit box)
-        if sample_transform is not None:
-            y_2_val = sample_transform(x_2, y_2)
+        if len(x_2) < min_points:
+            mask = np.ones([len(x_2) - 1], dtype=bool)
         else:
-            y_2_val = y_2
+            # represent the data as a path in N dimensions (scaled to unit box)
+            if sample_transform is not None:
+                y_2_val = sample_transform(x_2, y_2)
+            else:
+                y_2_val = y_2
 
-        p = np.r_['0',
-                  x_2[None, :],
-                  y_2_val.real.reshape(-1, y_2_val.shape[-1]),
-                  y_2_val.imag.reshape(-1, y_2_val.shape[-1])
-                  ]
+            p = np.r_['0',
+                      x_2[None, :],
+                      y_2_val.real.reshape(-1, y_2_val.shape[-1]),
+                      y_2_val.imag.reshape(-1, y_2_val.shape[-1])
+                      ]
 
-        sz = (p.shape[0] - 1) // 2
+            sz = (p.shape[0] - 1) // 2
 
-        xscale = x_2.ptp(axis=-1)
-        yscale = np.abs(y_2_val.ptp(axis=-1)).ravel()
+            xscale = x_2.ptp(axis=-1)
+            yscale = np.abs(y_2_val.ptp(axis=-1)).ravel()
 
-        p[0] /= xscale
+            p[0] /= xscale
 
-        p[1:sz + 1] /= yscale[:, None]
-        p[sz + 1:] /= yscale[:, None]
+            p[1:sz + 1] /= yscale[:, None]
+            p[sz + 1:] /= yscale[:, None]
 
-        # compute the length of each line segment in the path
-        dp = np.diff(p, axis=-1)
-        s = np.sqrt((dp**2).sum(axis=0))
-        s_tot = s.sum()
+            # compute the length of each line segment in the path
+            dp = np.diff(p, axis=-1)
+            s = np.sqrt((dp**2).sum(axis=0))
+            s_tot = s.sum()
 
-        # compute the angle between consecutive line segments
-        dp /= s
-        dcos = np.arccos(np.clip((dp[:, 1:] * dp[:, :-1]).sum(axis=0), -1, 1))
+            # compute the angle between consecutive line segments
+            dp /= s
+            dcos = np.arccos(np.clip((dp[:, 1:] * dp[:, :-1]).sum(axis=0), -1, 1))
 
-        # determine where to subdivide: the condition is roughly that
-        # the total length of the path (in the scaled data) is computed
-        # to accuracy `tol`
-        dp_piece = dcos * .5 * (s[1:] + s[:-1])
-        mask = (dp_piece > tol * s_tot)
+            # determine where to subdivide: the condition is roughly that
+            # the total length of the path (in the scaled data) is computed
+            # to accuracy `tol`
+            dp_piece = dcos * .5 * (s[1:] + s[:-1])
+            mask = (dp_piece > tol * s_tot)
 
-        mask = np.r_[mask, False]
-        mask[1:] |= mask[:-1].copy()
+            mask = np.r_[mask, False]
+            mask[1:] |= mask[:-1].copy()
 
-    # -- Refine, if necessary
+        # -- Refine, if necessary
 
-    if mask.any():
-        return _sample_function(func, x_2, y_2, mask, tol=tol, depth=depth + 1,
-                                min_points=min_points, max_level=max_level,
-                                sample_transform=sample_transform)
-    else:
-        return x_2, y_2
+        if mask.any():
+            return _sample_function(func, x_2, y_2, mask, tol=tol, depth=depth + 1,
+                                    min_points=min_points, max_level=max_level,
+                                    sample_transform=sample_transform)
+        else:
+            return x_2, y_2
